Update MyApps install script
Now install script can also be used for updating MyApps

+++ packaging/install.sh
@@ -1,12 +1,12 @@
 #!/bin/bash
 
-echo "MyApps installer"
+echo "MyApps installer/updater"
 echo "Run this as root or it won't work properly!"
 
 echo " -> Copying application"
+rm -rf /usr/bin/myapps
 cp myapps /usr/bin/myapps
 chmod a+x /usr/bin/myapps
-
 
 echo " -> Creating directories"
 if [ ! -d /usr/share/myapps ]; then
@@ -17,13 +17,20 @@ if [ ! -d /usr/share/myapps/repos ]; then
 fi
 
 echo " -> Creating files..."
-cat >/usr/share/myapps/config.yaml <<EOF
+if [ ! -e /usr/share/myapps/config.yaml ]; then
+  cat >/usr/share/myapps/config.yaml <<EOF
 default_repo: default
 EOF
-cat >/usr/share/myapps/deployments.yaml <<EOF
+fi
+
+if [ ! -e /usr/share/myapps/deployments.yaml ]; then
+  cat >/usr/share/myapps/deployments.yaml <<EOF
 []
 EOF
-cat >/usr/share/myapps/repos/default.yaml <<EOF
+fi
+
+if [ ! -e /usr/share/myapps/repos/default.yaml ]; then
+  cat >/usr/share/myapps/repos/default.yaml <<EOF
 name: default
 description: Default repository
 maintainer: MyApps
@@ -50,35 +57,7 @@ contents: []
 # - LATEST_VERSION (in install & update scenarios)
 # - CURRENT_VERSION (in update & uninstall scenario)
 EOF
-cat >/usr/share/myapps/repos/icikowski.yaml <<'EOF'
-name: icikowski
-description: Repository for my personal projects
-maintainer: Piotr Icikowski
-contents:
-  - name: gpts
-    description: General Purpose Test Service
-    install_scenario: 
-      - rm -rf /tmp/gpts-install
-      - mkdir /tmp/gpts-install
-      - wget -q -O /tmp/gpts-install/gpts https://github.com/Icikowski/GPTS/releases/download/v${LATEST_VERSION}/gpts-${LATEST_VERSION}-linux-amd64
-      - mv /tmp/gpts-install/gpts /usr/local/bin/gpts
-      - rm -rf /tmp/gpts-install
-      - chmod a+x /usr/local/bin/gpts
-    update_scenario:
-      - rm -f /usr/local/bin/gpts
-      - rm -rf /tmp/gpts-install
-      - mkdir /tmp/gpts-install
-      - wget -q -O /tmp/gpts-install/gpts https://github.com/Icikowski/GPTS/releases/download/v${LATEST_VERSION}/gpts-${LATEST_VERSION}-linux-amd64
-      - mv /tmp/gpts-install/gpts /usr/local/bin/gpts
-      - rm -rf /tmp/gpts-install
-      - chmod a+x /usr/local/bin/gpts
-    uninstall_scenario:
-      - rm -f /usr/local/bin/gpts
-    newest_version_check:
-      - curl -s https://charts.icikowski.pl/index.json | jq -r ".entries.gpts[0].version"
-    current_version_check:
-      - gpts --service-port -1 | head -n 1 | jq -r ".version"
-EOF
+fi
 
 echo " -> Enabling bash completion"
 cat >/etc/bash_completion.d/myapps <<'EOF'
@@ -107,8 +86,7 @@ echo " -> DONE"
 echo ""
 echo "You can change the default repository by editing file:"
 echo "  /usr/share/myapps/config.yaml"
-echo "You can check the repository structure in files:"
-echo "  /usr/share/myapps/repos/default.yaml"
-echo "  /usr/share/myapps/repos/icikowski.yaml"
+echo "Check for more pre-made repositories:"
+echo "  https://github.com/Icikowski/MyApps/tree/master/repositories"
 echo ""
 echo "Have fun!"
